+++ config.h
@@ -133,7 +133,7 @@
 #define PACKAGE_NAME "ttns-deck"
 
 /* Define to the full name and version of this package. */
-#define PACKAGE_STRING "ttns-deck 0.1.16-ttns-pre.3"
+#define PACKAGE_STRING "ttns-deck 0.1.16-ttns-pre.4"
 
 /* Define to the one symbol short name of this package. */
 #define PACKAGE_TARNAME "ttns-deck"
@@ -142,7 +142,7 @@
 #define PACKAGE_URL ""
 
 /* Define to the version of this package. */
-#define PACKAGE_VERSION "0.1.16-ttns-pre.3"
+#define PACKAGE_VERSION "0.1.16-ttns-pre.4"
 
 /* Define to 1 if all of the C89 standard headers exist (not just the ones
    required in a freestanding environment). This macro is provided for
@@ -150,7 +150,7 @@
 #define STDC_HEADERS 1
 
 /* Version number of package */
-#define VERSION "0.1.16-ttns-pre.3"
+#define VERSION "0.1.16-ttns-pre.4"
 
 /* Define to '__inline__' or '__inline' if that's what the C compiler
    calls it, or to nothing if 'inline' is not supported under any name.  */